+++ nengo_dl/tensor_graph.py
@@ -11,8 +11,7 @@ from nengo.exceptions import SimulationError
 from nengo.neurons import Direct
 import tensorflow as tf
 
-from nengo_dl import (builder, graph_optimizer, signals, utils, tensor_node,
-                      nengo_version)
+from nengo_dl import builder, graph_optimizer, signals, utils, tensor_node
 
 logger = logging.getLogger(__name__)
 
@@ -86,10 +85,7 @@ class TensorGraph(object):
         start = time.time()
 
         # group mergeable operators
-        if nengo_version < (2, 4, 0):
-            plan = graph_optimizer.greedy_planner(operators)
-        else:
-            plan = graph_optimizer.transitive_planner(operators)
+        plan = graph_optimizer.tree_planner(operators)
 
         # order signals/operators to promote contiguous reads
         sigs, self.plan = graph_optimizer.order_signals(plan, n_passes=10)